last

+++ src/simulation.py
@@ -1,5 +1,5 @@
 
 
 class Simulation:
-    def __init__(self, time_stamp : int) -> None:
+    def __init__(self, time_stamp : float) -> None:
         pass
